siphon off deposit-safety code+tests
